Alterando funcionamento da autenticação

+++ API-C#/WorkFlow.Domain/Identity/Role.cs
@@ -2,7 +2,7 @@ using Microsoft.AspNetCore.Identity;
 
 namespace WorkFlow.Domain.Identity
 {
-    public class Role : IdentityRole<long>
+    public class Role : IdentityRole<int>
     {
         public List<UserRole> UserRoles { get; set; }
     }
