+++ API-C#/WorkFlow.Domain/Identity/UserRole.cs
@@ -2,7 +2,7 @@ using Microsoft.AspNetCore.Identity;
 
 namespace WorkFlow.Domain.Identity
 {
-    public class UserRole : IdentityUserRole<long>
+    public class UserRole : IdentityUserRole<int>
     {
         public User User { get; set; }
         public Role Role { get; set; }
